Add the Package Support Explorer page
This is a simple client side AJAX app that parses the
support_info.xml file and matches a list of packages
supplied by the user to it.

This allows a copy&paste of "rpm -qa" and similar to get
a complete look at what package support statements are
applicable to that list of packages.