Color support statement by days remaining
Adds some client side javascript to display different colors
based on the number of days left of support.

We also incorporate this into the Package Support Explorer,
and fix a bug where the support statement list wasn't updated
when the date changed (thus the colors of the dates didn't
update).

Fixes: https://github.com/amazonlinux/al1-support-statements/issues/1